[#71] refactor: formatingDate함수 파일 위치 변경
